Move entity in configuration directory
